+++ smos/__init__.py
@@ -1,6 +1,1 @@
-import pkg_resources
-
-try:
-    __version__ = pkg_resources.get_distribution(__name__).version
-except:
-    __version__ = 'unknown'
+# -*- coding: utf-8 -*-