+++ CHANGES.rst
@@ -9,7 +9,7 @@ Version 0.x
 Version 0.1
 ===========
 
-- First version of SMOS IC image and TS reader
+- First version of L3 SMOS IC image and TS reader
 - Include time series reshuffling
 - Include tests and CI
 - Basic documentation